rename data-tiles/content pkg to data-tiles/grid
