+++ data-tiles/cmd/generate-breaks/main.go
@@ -9,21 +9,21 @@ import (
 	"path/filepath"
 
 	"github.com/ONSdigital/dp-geodata-api/data-tiles/cat"
-	"github.com/ONSdigital/dp-geodata-api/data-tiles/content"
 	"github.com/ONSdigital/dp-geodata-api/data-tiles/geos"
+	"github.com/ONSdigital/dp-geodata-api/data-tiles/grid"
 	"github.com/ONSdigital/dp-geodata-api/data-tiles/types"
 	"github.com/jtrim-ons/ckmeans/pkg/ckmeans"
 )
 
 func main() {
 	catfile := flag.String("c", "categories.txt", "text file holding list of categories to use")
-	contentfile := flag.String("q", "DataTileGrid.json", "json tile description file")
+	gridfile := flag.String("q", "DataTileGrid.json", "json tile description file")
 	geodir := flag.String("G", "data/processed/geo", "directory holding geojson files for each geotype")
 	metdir := flag.String("M", "data/processed/metrics", "directory holding metrics files for each category")
 	outdir := flag.String("O", "data/output/breaks", "output directory")
 	flag.Parse()
 
-	quads, err := content.Load(*contentfile)
+	quads, err := grid.Load(*gridfile)
 	if err != nil {
 		log.Fatal(err)
 	}
