+++ data-tiles/cmd/generate-tiles/main.go
@@ -5,13 +5,13 @@ import (
 	"log"
 
 	"github.com/ONSdigital/dp-geodata-api/data-tiles/cat"
-	"github.com/ONSdigital/dp-geodata-api/data-tiles/content"
 	"github.com/ONSdigital/dp-geodata-api/data-tiles/geos"
+	"github.com/ONSdigital/dp-geodata-api/data-tiles/grid"
 )
 
 func main() {
 	catfile := flag.String("c", "categories.txt", "text file holding list of categories to use")
-	contentfile := flag.String("q", "DataTileGrid.json", "json tile description file")
+	gridfile := flag.String("q", "DataTileGrid.json", "json tile description file")
 	geodir := flag.String("G", "data/processed/geo", "directory holding geojson files for each geotype")
 	metdir := flag.String("M", "data/processed/metrics", "directory holding metrics files for each category")
 	outdir := flag.String("O", "data/output/tiles", "output directory")
@@ -32,7 +32,7 @@ func main() {
 		log.Fatal(err)
 	}
 
-	quads, err := content.Load(*contentfile)
+	quads, err := grid.Load(*gridfile)
 	if err != nil {
 		log.Fatal(err)
 	}
